adding chroma db manager

+++ rag_llm.py
@@ -1,6 +1,8 @@
 import os
 import shutil
 
+import chromadb
+
 from langchain.text_splitter import RecursiveCharacterTextSplitter
 from langchain.schema import Document
 from langchain.vectorstores.chroma import Chroma
@@ -12,7 +14,6 @@ class rag_llm:
     
     def __init__(self) -> None:
         
-        # self.chroma_client = chromadb.Client()
         self.embeddings = OllamaEmbeddings(model="llama3")
         
         self.data_path = 'data'
@@ -25,20 +26,31 @@ class rag_llm:
         if os.path.exists(self.db_path):
             shutil.rmtree(self.db_path)
         os.makedirs(self.db_path)
-    
+        
+        self.get_chromadb()
+        
+        self.make_db_retriever()
+            
     def upload_data(self, uploaded_files):
         self.docs = []
         
         for file in uploaded_files:
-            file_path = os.path.join(self.data_path, file.name)
-            print("PATH :", file_path)
-            with open(file_path, 'wb') as f:
-                f.write(file.getvalue())
+            if not os.path.isfile(os.path.join("data", file.name)):
+            
+                file_path = os.path.join(self.data_path, file.name)
+                print("PATH :", file_path)
+                with open(file_path, 'wb') as f:
+                    f.write(file.getvalue())
+                    
+                loader = PyPDFLoader(file_path)
+                self.docs.extend(loader.load())
                 
-            loader = PyPDFLoader(file_path)
-            self.docs.extend(loader.load())
+        if self.docs:
+            self.add_documents()
+        
+        self.make_db_retriever()
                 
-    def make_vector_db(self):
+    def add_documents(self):
         
         chunks = self.split_text(self.docs)
         self.retriever = self.save_to_chroma(chunks)
@@ -55,25 +67,33 @@ class rag_llm:
         
         print(f"Split {len(documents)} documents into {len(chunks)} chunks.")
         
-        # print(chunks[0])
-
         return chunks
     
     def save_to_chroma(self, chunks):
         # Create a new DB from the documents.
-        db = Chroma.from_documents(
-            chunks, self.embeddings, persist_directory = self.db_path
-        )
+        self.db.add_documents(chunks)
+        
+        self.db.persist()
 
         print(f"Saved {len(chunks)} chunks to {self.db_path}.")
+        
+    def get_chromadb(self):
+        
+        if os.path.isfile("vec_db/chroma.sqlite3"):
+            self.db = Chroma(persist_directory=self.db_path, embedding_function=self.embeddings)
+        
+        else:
+            client = chromadb.PersistentClient()
+            collection = client.get_or_create_collection("Documents")
+            self.db = Chroma(client=client, collection_name="Documents", embedding_function=self.embeddings)
+        
+        print(self.db)
     
-    def get_data_retriever(self):
-        retriever = db.as_retriever(
+    def make_db_retriever(self):
+        self.retriever = self.db.as_retriever(
             search_type="similarity_score_threshold",
             search_kwargs={
                 "k": 3,
                 "score_threshold": 0.5,
             },
         )
-        
-        return retriever
